Fix and reorder test
